minor refactor and remove unused tests
