+++ src/mavetools/convert/__init__.py
@@ -1,4 +1,3 @@
-from mavetools.convert.codon import codon_sub_to_mavehgvs
 from mavetools.convert.enrich import seqid_to_variant
 
-__all__ = ["codon_sub_to_mavehgvs", "seqid_to_variant"]
+__all__ = ["seqid_to_variant"]
