+++ src/mavetools/util/codon.py
